Valid data rgb and texture path | Fix norme and clean archi
