+++ assets/map/map_error/map_error.cub
@@ -1,9 +1,9 @@
 NO ./assets/texture/NO.xpm
-SO ./assets/texture/SO.xpm
-WE ./assets/texture/WE.xpm
-EA ./assets/texture/EA.xpm
-F 220,100,0
-C 225,30,0
+SO assets/texture/SO.xpm
+WE assets/texture/WE.xpm
+EA assets/texture/EA.xpm
+F 220, 100,1
+C 225,30, 0
 
         1111111111111111111111111
         1000000000110000000000001
