+++ include/error.h
@@ -6,7 +6,7 @@
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/15 22:37:24 by mcogne--          #+#    #+#             */
-/*   Updated: 2025/01/19 23:45:30 by mcogne--         ###   ########.fr       */
+/*   Updated: 2025/01/20 21:47:24 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,7 @@ typedef enum e_error
 	ERR_PARSE_WALL,
 	ERR_PARSE_DATA,
 	ERR_PARSE_PLAYER,
+	ERR_MLX,
 	ERR_GAME
 }	t_error;
