+++ src/debug/debug_parsing.c
@@ -6,7 +6,7 @@
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/16 00:40:26 by mcogne--          #+#    #+#             */
-/*   Updated: 2025/01/19 22:53:30 by mcogne--         ###   ########.fr       */
+/*   Updated: 2025/01/20 21:33:21 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
