+++ src/parsing/validate/valid_player.c
@@ -1,25 +1,37 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exec.c                                             :+:      :+:    :+:   */
+/*   valid_player.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/17 04:31:58 by mcogne--          #+#    #+#             */
-/*   Updated: 2025/01/20 21:06:43 by mcogne--         ###   ########.fr       */
+/*   Created: 2025/01/20 20:31:48 by mcogne--          #+#    #+#             */
+/*   Updated: 2025/01/20 20:32:06 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-short	exec(t_env *env)
+short	valid_player(t_map *map, char **grid)
 {
-	if (ft_mlx_init_win(env, env->mlx))
-		return (env->err = ERR_MLX, 1);
-	if (ft_mlx_init_events(env))
-		return (env->err = ERR_MLX, 1);
-	if (render_map_2d(env->map, env->mlx))
-		return (env->err = ERR_GAME, 1);
-	mlx_loop(env->mlx->id);
+	size_t	count_player;
+	size_t	y;
+	size_t	x;
+
+	y = 1;
+	count_player = 0;
+	while (y < map->height)
+	{
+		x = 0;
+		while (grid[y][x])
+		{
+			if (ft_strchr(CARAC_PLAYER, grid[y][x]))
+				count_player++;
+			x++;
+		}
+		y++;
+	}
+	if (count_player != 1)
+		return (1);
 	return (0);
 }
