+++ src/utils/get_msg_error.c
@@ -6,7 +6,7 @@
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/19 18:34:44 by mcogne--          #+#    #+#             */
-/*   Updated: 2025/01/19 23:55:49 by mcogne--         ###   ########.fr       */
+/*   Updated: 2025/01/20 21:50:47 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,8 @@ char	*get_msg_error(size_t err)
 		return ("An unexpected error occurred.");
 	if (err == ERR_INIT_ENV)
 		return ("Error initializing data structures.");
+	if (err == ERR_PARSE_READ)
+		return ("Error occurred while read the file map.");
 	if (err == ERR_PARSE_PROCESS)
 		return ("Error occurred while loading the map.");
 	if (err == ERR_PARSE_CARAC)
@@ -25,10 +27,12 @@ char	*get_msg_error(size_t err)
 	if (err == ERR_PARSE_WALL)
 		return ("Invalid Map: Map is not enclosed by walls.");
 	if (err == ERR_PARSE_DATA)
-		return ("Error occurred while loading map data.");
+		return ("Error occurred while loading data (RGB or Path Textures).");
 	if (err == ERR_PARSE_PLAYER)
 		return ("Invalid Map: No or more than one players present in the map.");
+	if (err == ERR_MLX)
+		return ("Error occurred while mlx functions core.");
 	if (err == ERR_GAME)
-		return ("Error occurred during the game.");
+		return ("Error occurred during the game process.");
 	return ("An unexpected error occurred.");
 }
