+++ src/utils/init_env.c
@@ -6,11 +6,24 @@
 /*   By: mcogne-- <mcogne--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/15 22:36:10 by mcogne--          #+#    #+#             */
-/*   Updated: 2025/01/17 04:42:20 by mcogne--         ###   ########.fr       */
+/*   Updated: 2025/01/20 21:35:14 by mcogne--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
+
+static void	set_rgb_err(t_textures *textures)
+{
+	size_t	i;
+
+	i = 0;
+	while (i < 3)
+	{
+		textures->rgb_ceiling[i] = -1;
+		textures->rgb_floor[i] = -1;
+		i++;
+	}
+}
 
 short	init_env(t_env *env)
 {
@@ -33,5 +46,6 @@ short	init_env(t_env *env)
 	if (!env->textures)
 		return (1);
 	ft_memset(env->textures, 0, sizeof(t_textures));
+	set_rgb_err(env->textures);
 	return (0);
 }
